// src/api/generated/models/SubscriptionProvider.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Manager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.6.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const SubscriptionProvider = {
    Stripe: 'stripe',
    Apple: 'apple',
    GooglePlay: 'google_play'
} as const;
export type SubscriptionProvider = typeof SubscriptionProvider[keyof typeof SubscriptionProvider];


export function instanceOfSubscriptionProvider(value: any): boolean {
    for (const key in SubscriptionProvider) {
        if (Object.prototype.hasOwnProperty.call(SubscriptionProvider, key)) {
            if (SubscriptionProvider[key as keyof typeof SubscriptionProvider] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SubscriptionProviderFromJSON(json: any): SubscriptionProvider {
    return SubscriptionProviderFromJSONTyped(json, false);
}

export function SubscriptionProviderFromJSONTyped(json: any, ignoreDiscriminator: boolean): SubscriptionProvider {
    return json as SubscriptionProvider;
}

export function SubscriptionProviderToJSON(value?: SubscriptionProvider | null): any {
    return value as any;
}

export function SubscriptionProviderToJSONTyped(value: any, ignoreDiscriminator: boolean): SubscriptionProvider {
    return value as SubscriptionProvider;
}